ARM: i.MX: external NAND boot: make dtb boarddata work
If we are running from NFC SRAM and we are passed boarddata
containing a devicetree pointer then it point to an address relative
to the NFC SRAM start. First thing we do is to copy the initial
binary to SDRAM and jump there. The devicetree pointer has to be
adjusted by this offset.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>