images: fix image size in pblx
The pblx is a self extracting barebox binary. This doesn't have
the size of the image correctly set because the linker doesn't
generate it for relocatable binaries.
This currently only works on ARM, but this is the only architecture
supporting multi images anyway. TO make it work on other architectures
fix_size would have to be extended to recognize other images.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>