ARM: i.MX6 Sabrelite: Fix ethernet
The board code does a phy reset. This implicitly requests the phy reset
gpio. This gpio is also registered in the devicetree as phy reset gpio,
so the fec driver probe can't request the gpio and bails out with -EBUSY.
Fix this by freeing the phy reset gpio in the board code. While at it use
gpio_request_array for the gpios.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>