ARM: dts: Use local version of rk3188-clocks.dtsi
As of 3.17-rc1 the upstream dts sources for the rockchip rk3188 use
incompatible clock bindings. We currently do not have the resources
to update the clock driver to this new binding, so we use local copies
of the Linux 3.16 dtsi files. Due to the incompatible change the
internal device tree won't work with kernels newer than 3.17-rc1.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>